Add files via upload
bpm_detector_ros2.py is probably the only one we actually need, it has the bpm detection code in it. bpm_detector_launch I included so you could integrate the content from that launch file into our current launch file. The bpm_visualization_node will likely be helpful for debugging but impractical to use in conjunction with robot controls as it will likely take a lot of processing power and make everything slower. The readme walks through the topics and parameters for integration with the robot controls code so that you can subscribe to things that are being published.